requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/development.txt requirements/development.in
#
-e git+https://github.com/Saritasa/bravado.git#egg=bravado
alabaster==0.7.10         # via sphinx
amqp==2.2.2               # via kombu
apns2==0.3.0              # via django-push-notifications
argh==0.26.2              # via sphinx-autobuild, watchdog
asn1crypto==0.24.0        # via cryptography
autopep8==1.3.5
babel==2.5.3              # via flower, sphinx
backcall==0.1.0           # via ipython
bcrypt==3.1.4             # via paramiko
billiard==3.5.0.3         # via celery
bleach==2.1.3             # via nbconvert
blockdiag==1.5.3          # via sphinxcontrib-blockdiag
boltons==18.0.0
boto3==1.6.22
botocore==1.9.22          # via boto3, s3transfer
bravado-core==4.13.2
celery==4.1.0
certifi==2018.1.18        # via opbeat, requests
cffi==1.11.5              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coverage==4.5.1
cryptography==2.2.2       # via apns2, http-ece, paramiko, py-vapid, pywebpush
decorator==4.2.1          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid
django-allauth==0.35.0
django-appconf==1.0.2     # via django-imagekit
django-cacheops==4.0.6
django-celery-beat==1.1.1
django-constance[database]==2.2.0
django-cors-headers==2.2.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-extensions==2.0.6
django-filter==1.1.0
django-hstore==1.4.2
django-imagekit==4.0.2
django-jenkins==0.110.0
django-localflavor==2.0
django-model-utils==3.1.1
django-object-actions==1.0.0
django-pdb==0.6.1
django-picklefield==1.0.0
django-push-notifications==1.6.0
django-redis==4.9.0
django-rest-auth==0.9.3
django-storages==1.6.6
django-taggit-serializer==0.1.5
django-taggit==0.22.2
django-test-plus==1.0.22
django-test-without-migrations==0.6
django==1.11.8
djangorestframework-gis==0.12
djangorestframework-hstore==1.3
djangorestframework==3.7.7
docopt==0.6.2             # via ptpython
docutils==0.14            # via botocore, sphinx
drf-nested-routers==0.90.2
elpy==1.19.0
entrypoints==0.2.3        # via nbconvert
fabric3==1.14.post1
factory-boy==2.10.0
faker==0.8.12             # via factory-boy
first==2.0.1              # via pip-tools
flake8==3.5.0
flower==0.9.2
freezegun==0.3.10
funcparserlib==0.3.6      # via blockdiag
funcy==1.10.1
h2==2.6.2                 # via hyper
hpack==3.0.0              # via h2
html5lib==1.0.1           # via bleach
http-ece==1.0.5           # via pywebpush
hyper==0.7.0              # via apns2
hyperframe==3.2.0         # via h2, hyper
idna==2.6                 # via cryptography, requests
imagesize==1.0.0          # via sphinx
importmagic==0.1.7
inflection==0.3.1
ipdb==0.11
ipy==0.83
ipykernel==4.8.2          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0
ipython==6.3.0
ipywidgets==7.2.0         # via jupyter
isort==4.3.4
jedi==0.11.1
jinja2==2.10              # via nbconvert, notebook, sphinx
jmespath==0.9.3           # via boto3, botocore
jsonref==0.1              # via bravado-core
jsonschema[format]==2.6.0  # via bravado-core, nbformat, swagger-spec-validator
jupyter-client==5.2.3     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==5.2.0    # via jupyter
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
kombu==4.1.0              # via celery
livereload==2.5.1         # via sphinx-autobuild
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8, pylama
mistune==0.8.3            # via nbconvert
msgpack-python==0.5.6     # via bravado-core
munch==2.2.0
nbconvert==5.3.1          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.4.1           # via jupyter, widgetsnbextension
oauthlib==2.0.7           # via requests-oauthlib
opbeat==3.6.1
packaging==17.1           # via sphinx
pandocfilters==1.4.2      # via nbconvert
paramiko==2.4.1           # via fabric3
parso==0.1.1              # via jedi
pathtools==0.1.2          # via sphinx-autobuild, watchdog
pdfkit==0.6.1
pep8==1.7.1
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
pilkit==2.0
pillow==5.1.0             # via blockdiag
pip-tools==1.11.0
pockets==0.6.2            # via sphinxcontrib-napoleon
port_for==0.3.1           # via sphinx-autobuild
prompt-toolkit==1.0.15    # via ipython, jupyter-console, ptpython
psycopg2==2.7.4
ptipython==1.0.0
ptpython==0.41
ptyprocess==0.5.2         # via pexpect, terminado
py-vapid==1.3.1           # via pywebpush
pyasn1==0.4.2             # via paramiko
pycodestyle==2.3.1        # via autopep8, flake8, pylama
pycparser==2.18           # via cffi
pydocstyle==2.1.1         # via pylama
pyflakes==1.6.0           # via flake8, pylama
pygments==2.2.0           # via ipython, jupyter-console, nbconvert, ptpython, qtconsole, sphinx
pyjwt==1.6.1              # via apns2, twilio
pylama==7.4.3
pynacl==1.2.1             # via paramiko
pyparsing==2.2.0          # via packaging
pysocks==1.6.8            # via twilio
python-dateutil==2.6.1    # via botocore, bravado-core, faker, freezegun, jupyter-client
python-http-client==3.0.0  # via sendgrid
python3-openid==3.1.0     # via django-allauth
pytz==2018.3
pywebpush==1.7.0          # via django-push-notifications
pyyaml==3.12              # via bravado-core, sphinx-autobuild, watchdog
pyzmq==17.0.0             # via jupyter-client
qtconsole==4.3.1          # via jupyter
redis==2.10.6
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.4          # via django-allauth, pywebpush, requests-oauthlib, sphinx, twilio
rfc3987==1.3.7            # via jsonschema
rope==0.10.7
s3transfer==0.1.13        # via boto3
send2trash==1.5.0         # via notebook
sendgrid-django==4.2.0
sendgrid==3.6.5           # via sendgrid-django
simplegeneric==0.8.1      # via ipython
simplejson==3.13.2        # via bravado-core
six==1.11.0               # via bcrypt, bleach, bravado-core, cryptography, django-cacheops, django-coverage-plugin, django-extensions, django-imagekit, django-rest-auth, djangorestframework-gis, fabric3, faker, freezegun, html5lib, livereload, munch, packaging, pip-tools, pockets, prompt-toolkit, pydocstyle, pynacl, python-dateutil, sphinx, sphinxcontrib-napoleon, swagger-spec-validator, traitlets, twilio
snowballstemmer==1.2.1    # via pydocstyle, sphinx
sphinx-autobuild==0.7.1
sphinx-py3doc-enhanced-theme==2.4.0
sphinx-rtd-theme==0.2.4
sphinx==1.7.2             # via sphinxcontrib-blockdiag
sphinxcontrib-blockdiag==1.5.5
sphinxcontrib-napoleon==0.6.1
sphinxcontrib-websupport==1.0.1  # via sphinx
sqlparse==0.2.4           # via django-debug-toolbar
strict-rfc3339==0.7       # via jsonschema
swagger-spec-validator==2.1.0  # via bravado-core
tblib==1.3.2
terminado==0.8.1          # via notebook
testpath==0.3.1           # via nbconvert
text-unidecode==1.2       # via faker
tornado==5.0.1            # via flower, ipykernel, jupyter-client, livereload, notebook, sphinx-autobuild, terminado
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
twilio==6.11.0
urllib3==1.22             # via opbeat, requests
vine==1.1.4               # via amqp
watchdog==0.8.3           # via sphinx-autobuild
wcwidth==0.1.7            # via prompt-toolkit
webcolors==1.8.1          # via blockdiag, jsonschema
webencodings==0.5.1       # via html5lib
werkzeug==0.14.1
widgetsnbextension==3.2.0  # via ipywidgets
yapf==0.21.0
yelp-bytes==0.3.0
yelp-encodings==0.1.3     # via yelp-bytes
